requirements.txt
aiohttp
aiosignal
alembic
annotated-types
anyio
async-timeout
asyncpg
attrs
black
certifi
charset-normalizer
click
dnspython
ecdsa
email_validator
exceptiongroup
fastapi
fastapi-cli
frozenlist
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
Mako
markdown-it-py
MarkupSafe
mdurl
multidict
mypy-extensions
orjson
packaging
passlib
pathspec
pillow
platformdirs
pyasn1
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
PyJWT
python-dotenv
python-jose
python-multipart
PyYAML
requests
rich
rsa
shellingham
six
sniffio
SQLAlchemy
starlette
tomli
typer
typing_extensions
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
yarl